requirements-hiervae.txt
about-time==4.2.1
aim==3.17.5
aim-ui==3.17.5
aimrecords==0.0.7
aimrocks==0.4.0
aiofiles==23.2.1
alembic==1.12.0
alive-progress==3.1.4
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.4.0
autograd==1.6.2
backcall==0.2.0
backoff==2.2.1
base58==2.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cma==3.2.2
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.1
filelock==3.12.4
fonttools==4.42.1
future==0.18.3
grapheme==0.6.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipython==8.12.2
jedi==0.19.0
Jinja2==3.1.2
kiwisolver==1.4.5
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
monotonic==1.6
networkx==2.6.3
numpy==1.20.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py3nvml==0.2.7
pycparser==2.21
pydantic==2.4.0
pydantic_core==2.10.0
Pygments==2.16.1
pymoo==0.6.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
rdkit-pypi==2020.9.5.2
requests==2.31.0
RestrictedPython==6.2
scipy==1.10.1
segment-analytics-python==2.2.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
stack-data==0.6.2
starlette==0.27.0
torch==1.13.1
tqdm==4.66.1
traitlets==5.10.0
typing_extensions==4.7.1
urllib3==2.0.5
uvicorn==0.23.2
wcwidth==0.2.6
wrapt==1.15.0
xmltodict==0.13.0
zipp==3.17.0
